requirements.txt
aiobotocore==2.24.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aioitertools==0.12.0
aiosignal==1.4.0
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
billiard==4.2.1
blinker==1.9.0
boto3==1.39.11
botocore==1.39.11
cachetools==6.1.0
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.9
cryptography==45.0.6
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.24.1
dvc==3.62.0
dvc-data==3.16.11
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-s3==3.2.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
filelock==3.19.1
flatten-dict==0.4.2
flufl.lock==8.2.0
frozenlist==1.7.0
fsspec==2025.7.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.45
grandalf==0.8
gto==1.7.2
hydra-core==1.3.2
idna==3.10
iterative-telemetry==0.0.10
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.4
narwhals==2.1.2
networkx==3.4.2
numpy==2.3.2
omegaconf==2.3.0
orjson==3.11.2
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pydot==4.0.1
pygit2==1.18.2
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rich==14.1.0
rpds-py==0.27.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
s3fs==2025.7.0
s3transfer==0.13.1
scmrepo==3.5.2
semver==3.0.4
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sqltrie==0.11.2
streamlit==1.48.1
tabulate==0.9.0
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
wrapt==1.17.3
yarl==1.20.1
zc.lockfile==3.0.post1

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-html==4.1.1

# Code quality tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
bandit==1.7.5
mypy==1.7.1
